requirements.txt
Django==1.11.2
django-cors-headers==2.1.0
django-filter==1.0.4
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
Markdown==2.6.8
psycopg2==2.7.1
PyJWT==1.5.2
pytz==2017.2
